Changed serializer and update read me
Read me now has now been updated. I kept the majority of the
crud stuff adding on the important parts for my own code.

Not sure if that is kosher, but I didn't want to reinvent the wheel.

Serializer for users now doesn't include email as that isn't nescessary
for anything on the front.